rename gpu_cpu to gpu_dram
